+++ CHANGELOG.md
@@ -6,11 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - YYYY-MM-DD
+### Added
+
+### Fixed
+
+### Changed
+
+### Removed
+
+### Deprecated
+
+## [0.1.8] - 2023-06-05
 
 ### Added
 
 - @imtbl/blockchain-data: Package for exposing blockchain data information.
-- @imtbl/passport: gardian domain in configuration
+- @imtbl/passport: guardian domain in configuration
 
 ### Fixed